requirements.txt
aiohttp>=3.9.1
psutil>=5.9.6
requests>=2.31.0
python-dotenv>=1.0.0
PyQt6>=6.6.1
rich>=13.7.0
python-telegram-bot>=20.7
pyyaml>=6.0.1
geoip2>=4.8.0
qrcode>=7.4.2
pillow>=10.0.0
